refactor(domain): add domain range prop to axis spec
