+++ src/lib/series/domains/x_domain.test.ts
@@ -590,12 +590,12 @@ describe('X Domain', () => {
     specDataSeries.set(ds1.id, ds1);
     specDataSeries.set(ds2.id, ds2);
     const { xValues } = getSplittedSeries(specDataSeries);
+
     const mergedDomain = mergeXDomain(
       [
         {
           seriesType: 'area',
           xScaleType: ScaleType.Linear,
-          xDomain: [0, 10],
         },
         {
           seriesType: 'line',
