+++ stories/line_chart.tsx
@@ -261,3 +261,4 @@ storiesOf('Line Chart', module)
       </Chart>
     );
   });
+
